test: implement base test case for reference endpoints and refactor existing tests
